class exercise about FreeRTOS
Just a simple exercise that I was doing while in class the program creates two tasks, taskA and taskB, which communicate with each other using a queue.

The logic of the queue:
The program creates a queue using xQueueCreate() and initializes it with a maximum length of 10 elements, each of size sizeof(char*).

Afterwards, the program starts the FreeRTOS scheduler using vTaskStartScheduler(), which will run the tasks in a cooperative multitasking environment until the program terminates.