2.1 Counter exercise
2.1.1 1 counter
Synchronization can be implemented in Threads in two ways: using the `synchronized` keyword with a method or using a `synchronized` block.

When the `synchronized` keyword is used with a method it affect the entire method so is similar to when we use multi-line comments.

When using synchronized block a single variable or shared resource within a method is marked to be protected so is like we are using a single line comment, you will want to use that when you have only one shared resource that you need to protect as it needs collaboration from the threads in my case only one variable had to be marked as synchronized(lock) inside the block