2.1.2 2 counters
This exercise is essentially the same as the one with 1 counter but this time it has 2 counters
I have created countDooku and countDracule
Each counter has its own lock object for syncronization as lockDooku and lockDracula
The `incrementDooku()` and `incrementDracula()` methods each increment their respective counter in
a synchronized block, ensuring that only one thread can increment the counter at a time.
The countIncrementer class than implements the Runnable and increments both counters 1 million
times in its run() method
The Start class finally creates a Counter with two CountIncrementer threads that operate the same
Counter, after both threads have completed the final counts are printed out. The output is as expected
2 million for each count.