2.4 wait() and notifyAll()
Implement the following system implementing a burger bar with customers and employees. Employees are making burgers and customers are eating burgers

what I learned from doing this exercise are the new concepts of wait() and notify All()
**Inter-thread communication: When the burger bar is full, employees have to wait until a customer eats a burger. When there are no burgers, customers have to wait until an employee makes a burger. The `wait()`** method is used to make a thread wait, and notifyAll()`** is used to wake up all waiting threads when the state of the burger bar changes.

Control of thread execution: By using Thread.sleep()`, you can simulate delays, such as the time it takes to make or eat a burger.